extracted logic from NotificationsUpdateService to NotificationsUpdateLogic for reuse
